debut du travail sur les ennemies et les combats

+++ src/Weapon.cpp
@@ -4,7 +4,7 @@ Weapon::~Weapon() {
     std::cout << "Dropped this useless sword" << std::endl;
 }
 
-Weapon::Weapon(const std::string &name, const std::string &type, int dmg, int range, int as, int hands) : _name(name), _dmg(dmg), _range(range), _atk_speed(as), _hands(hands) {}
+Weapon::Weapon(const std::string &name, const std::string &type, int dmg, float range, int as, int hands) : _name(name), _dmg(dmg), _range(range), _atk_speed(as), _hands(hands) {}
 
 std::string Weapon::getName() const {
     return _name;
@@ -18,7 +18,7 @@ int Weapon::getDmg() const {
     return _dmg;
 }
 
-int Weapon::getRange() const {
+float Weapon::getRange() const {
     return _range;
 }
 
@@ -34,7 +34,7 @@ void Weapon::setDmg(const int dmg) {
     this->_dmg = dmg;
 }
 
-void Weapon::setRange(const int range) {
+void Weapon::setRange(const float range) {
     this->_range = range;
 }
